+++ app/jobs/fan_out_printer_polls_job.rb
@@ -2,8 +2,8 @@ class FanOutPrinterPollsJob < ApplicationJob
   queue_as :default
 
   def perform
-    Printer.where.not(prusalink_key: nil).find_each do |printer|
-      PrinterPollJob.perform_later(printer.id)
+    Printer.find_each do |printer|
+      PrinterPollJob.perform_later(printer.id) if printer.prusalink?
     end
   end
 end